+++ src/jsp/com/cubrid/plcsql/handler/TestMain.java
@@ -37,8 +37,8 @@ import com.cubrid.plcsql.compiler.ParseTreeConverter;
 import com.cubrid.plcsql.compiler.ParseTreePrinter;
 import com.cubrid.plcsql.compiler.PcsLexerEx;
 import com.cubrid.plcsql.compiler.SemanticError;
-import com.cubrid.plcsql.compiler.ServerAPI;
-import com.cubrid.plcsql.compiler.SqlSemantics;
+import com.cubrid.plcsql.compiler.serverapi.ServerAPI;
+import com.cubrid.plcsql.compiler.serverapi.SqlSemantics;
 import com.cubrid.plcsql.compiler.StaticSqlCollector;
 import com.cubrid.plcsql.compiler.SyntaxError;
 import com.cubrid.plcsql.compiler.antlrgen.PcsParser;
